minor glitches, delete entities

+++ take-home/src/apartments/apartments.controller.ts
@@ -1,4 +1,17 @@
-import {Body, ClassSerializerInterceptor, Controller, Delete, Get, Param, Patch, Post, Query, Req, UseInterceptors} from '@nestjs/common';
+import {
+  Body,
+  ClassSerializerInterceptor,
+  Controller,
+  Delete,
+  Get,
+  HttpException, HttpStatus,
+  Param,
+  Patch,
+  Post,
+  Query,
+  Req,
+  UseInterceptors
+} from '@nestjs/common';
 import {ApartmentsService} from './apartments.service';
 import {CreateApartmentDto} from './dto/create-apartment.dto';
 import {UpdateApartmentDto} from './dto/update-apartment.dto';
@@ -7,6 +20,7 @@ import {Roles} from '../auth/auth.decorator';
 import {QueryApartmentDto} from './dto/query-apartment.dto';
 import {Apartment} from './entities/apartment.entity';
 import {serialize} from 'class-transformer';
+import {throwError} from 'rxjs';
 
 
 @Controller('apartments')
@@ -18,19 +32,20 @@ export class ApartmentsController {
   @Roles(Role.REALTOR, Role.ADMIN)
   create(@Body() createApartmentDto: CreateApartmentDto) {
     createApartmentDto.createdDate = new Date();
+    createApartmentDto.rentable = true;
     return this.apartmentsService.create(createApartmentDto);
   }
 
   @Get()
   async findAll(@Req() {user}) {
-    const apartments: Apartment[] = await this.apartmentsService.findAll(user.role);
+    const apartments: Apartment[] = await this.apartmentsService.findAll(user.role, user.id);
     return serialize(apartments);
   }
 
   @Get('findByFilter')
   async findByFilter(@Req() {user}, @Query() query) {
     let queryDto = new QueryApartmentDto(query);
-    const apartments: Apartment[] = await this.apartmentsService.findAll(user.role, queryDto.conditions());
+    const apartments: Apartment[] = await this.apartmentsService.findAll(user.role, user.id, queryDto.conditions());
     return serialize(apartments);
   }
 
@@ -38,18 +53,34 @@ export class ApartmentsController {
   @Get(':id')
   @UseInterceptors(ClassSerializerInterceptor)
   findOne(@Req() {user}, @Param('id') id: string) {
-    return this.apartmentsService.findOne(+id, user.role, {});
+    return this.apartmentsService.findOne(+id, user.role, user.id);
   }
 
   @Patch(':id')
   @Roles(Role.REALTOR, Role.ADMIN)
-  update(@Param('id') id: string, @Body() updateApartmentDto: UpdateApartmentDto) {
-    return this.apartmentsService.update(+id, updateApartmentDto);
+  async update(@Req() {user}, @Param('id') id: string, @Body() updateApartmentDto: UpdateApartmentDto) {
+    const apartment = await this.apartmentsService.findOne(+id, user.role, user.id)
+    if (apartment) {
+      return this.apartmentsService.update(+id, updateApartmentDto);
+    } else {
+      throw new HttpException({
+        status: HttpStatus.FORBIDDEN,
+        error: 'No Acceess',
+      }, HttpStatus.FORBIDDEN);
+    }
   }
 
   @Delete(':id')
   @Roles(Role.REALTOR, Role.ADMIN)
-  remove(@Param('id') id: string) {
-    return this.apartmentsService.remove(+id);
+  async remove(@Req() {user}, @Param('id') id: string) {
+    const apartment = await this.apartmentsService.findOne(+id, user.role, user.id)
+    if (apartment) {
+      return this.apartmentsService.remove(+id);
+    } else {
+      throw new HttpException({
+        status: HttpStatus.FORBIDDEN,
+        error: 'No Acceess',
+      }, HttpStatus.FORBIDDEN);
+    }
   }
 }
